+++ Tests/UnitTests/Core/Geometry/TrackingVolumeTests.cpp
@@ -21,10 +21,10 @@
 #include "Acts/Geometry/TrackingVolume.hpp"
 #include "Acts/Propagator/ActionList.hpp"
 #include "Acts/Propagator/ConstrainedStep.hpp"
+#include "Acts/Propagator/DebugOutputActor.hpp"
 #include "Acts/Propagator/Propagator.hpp"
+#include "Acts/Propagator/StandardAborters.hpp"
 #include "Acts/Propagator/StraightLineStepper.hpp"
-#include "Acts/Propagator/detail/DebugOutputActor.hpp"
-#include "Acts/Propagator/detail/StandardAborters.hpp"
 #include "Acts/Propagator/detail/SteppingLogger.hpp"
 
 #include "Acts/Geometry/GeometryContext.hpp"
