+++ Tests/UnitTests/Core/Material/VolumeMaterialMapperTests.cpp
@@ -21,11 +21,11 @@
 #include "Acts/Material/Material.hpp"
 #include "Acts/Material/MaterialMapUtils.hpp"
 #include "Acts/Material/VolumeMaterialMapper.hpp"
+#include "Acts/Propagator/DebugOutputActor.hpp"
 #include "Acts/Propagator/Navigator.hpp"
 #include "Acts/Propagator/Propagator.hpp"
+#include "Acts/Propagator/StandardAborters.hpp"
 #include "Acts/Propagator/StraightLineStepper.hpp"
-#include "Acts/Propagator/detail/DebugOutputActor.hpp"
-#include "Acts/Propagator/detail/StandardAborters.hpp"
 #include "Acts/Tests/CommonHelpers/FloatComparisons.hpp"
 #include "Acts/Utilities/Definitions.hpp"
 #include "Acts/Utilities/Helpers.hpp"
@@ -309,11 +309,11 @@ BOOST_AUTO_TEST_CASE(VolumeMaterialMapper_comparison_tests) {
 
   MagneticFieldContext mc;
 
-  using DebugOutput = Acts::detail::DebugOutputActor;
+  using DebugOutput = Acts::DebugOutputActor;
 
   // Launch propagation and gather result
   PropagatorOptions<ActionList<MaterialCollector, DebugOutput>,
-                    AbortList<detail::EndOfWorldReached>>
+                    AbortList<EndOfWorldReached>>
       po(gc, mc);
   po.maxStepSize = 1._mm;
   po.maxSteps = 1e6;
